chapter02/basic.py
import openai
import os

from dotenv import load_dotenv

load_dotenv()

openai.api_key = os.getenv("OPENAI_API_KEY")


def chat(message: str) -> str:
    response = openai.chat.completions.create(
        model="gpt-3.5-turbo",
        messages=[
            {"role": "system", "content": "You are a helpful assistant."},
            {"role": "user", "content": message},
        ],
        max_tokens=300,
        temperature=0.9,
    )

    return response.choices[0].message.content


if __name__ == "__main__":
    print(chat("What is the meaning of life?"))
